refactor: restructure members module procedures and fix type imports
  - Move members procedures into organized subdirectories (get-one/, get-many/)
  - Fix broken schema imports in procedures after file reorganization
  - Update type imports in UI components to use proper schema types
  - Remove dependencies on deleted members/schema.ts and members/types.ts files
  - Ensure proper TypeScript type inference for TRPC queries
