+++ src/modules/iracing/server/procedures/index.ts
@@ -1,4 +1,5 @@
 import { createTRPCRouter } from "@/trpc/init";
+
 import { getUserProcedure } from "./get-user";
 import { getUserRecentRacesProcedure } from "./get-user-recent-races";
 import { getUserSummaryProcedure } from "./get-user-summary";
