+++ src/modules/members/server/procedures/get-many/schema.ts
@@ -1,0 +1,24 @@
+import z from "zod";
+
+import { inferRouterOutputs } from "@trpc/server";
+import { AppRouter } from "@/trpc/routers/_app";
+
+import {
+  DEFAULT_PAGE,
+  DEFAULT_PAGE_SIZE,
+  MAX_PAGE_SIZE,
+  MIN_PAGE_SIZE,
+} from "@/modules/iracing/constants";
+
+export const GetMembersInputSchema = z.object({
+  page: z.number().default(DEFAULT_PAGE),
+  pageSize: z
+    .number()
+    .min(MIN_PAGE_SIZE)
+    .max(MAX_PAGE_SIZE)
+    .default(DEFAULT_PAGE_SIZE),
+    
+  search: z.string().nullish(),
+});
+
+export type GetMany = inferRouterOutputs<AppRouter>["members"]["getMany"];
