--- src/modules/members/types.ts
@@ -1,7 +1,0 @@
-import { inferRouterOutputs } from "@trpc/server";
-
-import { AppRouter } from "@/trpc/routers/_app";
-
-export type MembersGetOne = inferRouterOutputs<AppRouter>["members"]["getOne"];
-export type MembersGetMany =
-  inferRouterOutputs<AppRouter>["members"]["getMany"];
