+++ src/trpc/routers/_app.ts
@@ -12,5 +12,4 @@ export const appRouter = createTRPCRouter({
   iracing: iracingRouter,
 });
 
-// export type definition of API
 export type AppRouter = typeof appRouter;
